backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
PyMuPDF==1.25.4
python-docx==1.1.2
huggingface_hub==0.30.2
langchain==0.3.24
langgraph==0.4.5
langsmith==0.3.42
